requirements.txt
pandas>=2.0
openpyxl>=3.1
requests>=2.31
pyodbc>=5.0
python-dotenv>=1.0
pywin32>=306       # win32com para Pivot Tables (Windows only)
pytest>=7.4        # testes
